refactor: Boards 생성
